requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
attrs==23.2.0
frozenlist==1.4.1
idna==3.7
iso8601==1.1.0
multidict==6.0.5
py-cord==2.5.0
pydantic==2.7.3
pydantic_core==2.18.4
pypika-tortoise==0.1.6
pytz==2024.1
tortoise-orm==0.21.3
typing_extensions==4.12.2
yarl==1.9.4
python-dotenv==1.0.1
